menial fixes

+++ code/UI.py
@@ -32,19 +32,24 @@ def ui() -> bool:
                 print('No movies found')
         case 2:
             Film.add_by_api(input('Enter movie title: '))
+            Film.create_backup()
             print('Movie added')
         case 3:
-            for film in Film.all():
-                print(film.__repr__())
-            try:
-                user_input = int(input('Choose a movie: '))
-                if user_input not in [film.id for film in Film.all()]:
-                    print('Invalid input')
-                    return True
-                Film.get(user_input).delete()
-                print('Movie deleted')
-            except ValueError:
-                print('I need a number')
+            if Film.all():
+                for film in Film.all():
+                    print(film.__repr__())
+                try:
+                    user_input = int(input('Choose a movie: '))
+                    if user_input not in [film.id for film in Film.all()]:
+                        print('Invalid input')
+                        return True
+                    Film.get(user_input).delete()
+                    Film.create_backup()
+                    print('Movie deleted')
+                except ValueError:
+                    print('I need a number')
+            else:
+                print('No movies to delete')
         case _:
             print('I didn\'t get your input')
     return True
